kvserver: put snapshot strategy into its own file
Epic: CRDB-46488
Release note: none
