+++ pkg/kv/kvserver/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
         "flow_control_replica.go",
         "flow_control_replica_integration.go",
         "flow_control_stores.go",
+        "kv_snapshot_strategy.go",
         "lease_history.go",
         "lease_queue.go",
         "markers.go",
